refactor : rename SortByAdapter to SortingParameterAdapter
